FlexType: Make all the types generic between xtream providers
