+++ go.mod
@@ -1,0 +1,11 @@
+module github.com/tellytv/go.xtream-codes
+
+go 1.17
+
+require github.com/onsi/gomega v1.17.0
+
+require (
+	golang.org/x/net v0.0.0-20210428140749-89ef3d95e781 // indirect
+	golang.org/x/text v0.3.6 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+)
